Add a secure downloading system
This system allows for hotlink prevention and useage of ads (with a direct link)
To prevent hotlinking together, you can disable indexing of your directories
and then block all attempts to access any file thats not in resources/ or that is not index.php getdownload.php or download.php.
With these changes, the md5 is also shown at download time
and a ten second counter starts for the download.