Fix bug where up directory breaks resource CSS
The current way the code is, if index.php is invoked (like by pressing the up directory icon) you will get invalid pages. This is because the CSS path is trying to go through index.php, and not through the normal routes. To prevent this, simply add a forward slash in and the path becomes absolute based on the URI.